Remove compilation warnings
